fix bug run 1
