+++ communitymanagement/src/main/java/communitymanagement/dao/TicketCommentDao.java
@@ -9,11 +9,9 @@ import javax.persistence.criteria.Root;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import communitymanagement.model.Ticket;
 import communitymanagement.model.TicketComment;
 
 @Repository
