+++ communitymanagement/src/main/java/communitymanagement/model/Authorities.java
@@ -24,7 +24,7 @@ public class Authorities implements Serializable {
 
 	private int userId;
 	
-	private String userName;
+	private String username;
 
 	private String authorities;
 }
