+++ communitymanagement/src/main/java/communitymanagement/service/IssueCategoryService.java
@@ -2,10 +2,12 @@ package communitymanagement.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired; 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import communitymanagement.dao.IssueCategoryDao;
+import communitymanagement.dao.IssueDao;
+import communitymanagement.dao.LocationDao;
 import communitymanagement.model.Issue;
 import communitymanagement.model.IssueCategory;
 import communitymanagement.model.Location;
@@ -15,6 +17,12 @@ public class IssueCategoryService {
 
 	@Autowired
 	private IssueCategoryDao issueCategoryDao;
+	
+	@Autowired
+	private IssueDao issueDao;
+	
+	@Autowired
+	private LocationDao locationDao;
 	
 	public void addIssueCategory(IssueCategory issueCategory) {
 		issueCategoryDao.addIssueCategory(issueCategory);
@@ -37,6 +45,23 @@ public class IssueCategoryService {
 	}
 	
 	public void addIssueCategoryByName(String issueName, String locationName) {
-		issueCategoryDao.addIssueCategoryByName(issueName, locationName);
+		// issue
+		Issue issue = issueDao.getIssueByName(issueName);
+		if (issue == null) {
+			issue = new Issue();
+			issue.setIssueType(issueName);
+			issueDao.addIssue(issue);
+		}
+
+		// category
+		Location location = locationDao.getLocationByName(locationName);
+		if (location == null) {
+			location = new Location();
+			location.setLocationType(locationName);
+			locationDao.addLocation(location);
+		}
+
+		// issue category
+		addIssueCategory(issue, location);
 	}
 }
